messages view

+++ app/src/main/java/com/ekdorn/silentium/models/Message.kt
@@ -3,4 +3,4 @@ package com.ekdorn.silentium.models
 import com.ekdorn.silentium.core.Myte
 
 
-data class Message(val text: Myte, val date: Int, var read: Boolean, val author: Contact)
+data class Message(val text: Myte, val date: Long, var read: Boolean, val author: Contact)
